Refactor: hooks와 atom을 모듈화하여 관리하고 page단에서 모든 로직이 작동하도록 리팩토링
