+++ src/app/stage/page.tsx
@@ -1,21 +1,68 @@
 "use client";
 import React from "react";
-import { atom, useRecoilState, useRecoilValue } from "recoil";
+import { useEffect } from "react";
+import { useRecoilState, useRecoilValue, useSetRecoilState } from "recoil";
+import { styled } from "styled-components";
+import { useRouter } from "next/navigation";
+import { gptResultState, loadingState, stageNumberState, stageResultState } from "../atoms/atom";
+import { useStageNumberMemo } from "../hooks/hooks";
 import { ProgressBar } from "@/components/progressBar/ProgressBar";
 import { GameDescBox } from "@/components/GameDesc/GameDescBox";
-import { styled } from "styled-components";
+import { Loading } from "@/components/loading/Loading";
 import questions from "../../question.json";
-import { stageNumberState } from "../../components/GameDesc/GameDescBox";
+
+type StageResult = {
+  [key: string]: number;
+};
 
 export default function StagePage() {
-  const stageNumber = useRecoilValue<number>(stageNumberState);
+  const router = useRouter();
+  const setGptResult = useSetRecoilState(gptResultState);
+  const stageResult = useRecoilValue<StageResult>(stageResultState);
+  const [stageNumber, setStageNumber] = useRecoilState<number>(stageNumberState);
+  const [loadingOpen, setLoadingOepn] = useRecoilState<boolean>(loadingState);
   const { question, choices } = stageNumber === 11 ? { question: undefined, choices: undefined } : questions[stageNumber - 1];
 
+  async function clickHandlerGPT() {
+    try {
+      setLoadingOepn(true);
+      const response = await fetch("/api/generate", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({ value: stageResult }),
+      });
+
+      let data = await response.json();
+      if (response.status !== 200) {
+        throw data.error || new Error(`Request failed with status ${response.status}`);
+      }
+
+      if (typeof data === "string") {
+        data = JSON.parse(data);
+      }
+      setGptResult(data);
+      if (response.status === 200) {
+        router.push("/result");
+        setLoadingOepn(false);
+      }
+    } catch (error: any) {
+      console.error(error);
+      alert(error.message);
+    }
+  }
+
+  useStageNumberMemo(stageResult, clickHandlerGPT);
+
   return (
-    <DescWrapper>
-      <ProgressBar value={Number(stageNumber) * 10} />
-      <GameDescBox descHeader={`stage${stageNumber}`} desc={question} startButtonDesc={""} buttonDesc={choices} />
-    </DescWrapper>
+    <>
+      <DescWrapper>
+        <ProgressBar value={Number(stageNumber) * 10} />
+        <GameDescBox descHeader={`Stage${stageNumber}`} desc={question} startButtonDesc={""} buttonDesc={choices} />
+      </DescWrapper>
+      {loadingOpen && <Loading />}
+    </>
   );
 }
